operator button box stuff (not finished)

+++ src/OperatorController/LED.cpp
@@ -1,3 +1,7 @@
+/// \copyright Copyright (c) Argos FRC Team 1756.
+///            Open Source Software; you can modify and/or share it under the terms of
+///            the license file in the root directory of this project.
+
 ////////////////////////////////////////////////////////////////////////////////
 /// @file LED.cpp
 ///
@@ -24,36 +28,26 @@
 bool LED::GAMMA_INITIALIZED = false;
 uint8_t LED::GAMMA[256];
 
-LED::LED( uint8_t controlPin) : LED( controlPin,
-                                     (bool)false )
-{}
+LED::LED(uint8_t controlPin) : LED(controlPin, (bool)false) {}
 
-LED::LED( uint8_t controlPin,
-          bool    pwmEnable ) : LED( controlPin,
-                                     NOT_A_PIN,
-                                     pwmEnable )
-{}
+LED::LED(uint8_t controlPin, bool pwmEnable) : LED(controlPin, NOT_A_PIN, pwmEnable) {}
 
-LED::LED( uint8_t controlPin,
-          uint8_t groundPin,
-          bool    pwmEnable ) : m_controlPin(controlPin),
-                                m_groundPin(groundPin),
-                                m_pwmEnable(pwmEnable),
-                                m_initialized(false),
-                                m_pattern(LEDPattern::SOLID),
-                                m_illuminate(false)
-{
-  if(pwmEnable && !GAMMA_INITIALIZED)
-  {
+LED::LED(uint8_t controlPin, uint8_t groundPin, bool pwmEnable)
+    : m_controlPin(controlPin)
+    , m_groundPin(groundPin)
+    , m_pwmEnable(pwmEnable)
+    , m_initialized(false)
+    , m_pattern(LEDPattern::SOLID)
+    , m_illuminate(false) {
+  if (pwmEnable && !GAMMA_INITIALIZED) {
     GAMMA_INITIALIZED = true;
     // Populate Gamma correction lookup table
     // Gamma correction for x is:
     // ( x / MAXVAL )^2.5 * MAXVAL
-    for(int i = 0; i < 256; i++)
-    {
+    for (int i = 0; i < 256; i++) {
       float x = i;
       x /= 255;
-      x = pow(x, 2.5);
+      x = std::pow(x, 2.5);
       x *= 255;
 
       GAMMA[i] = x;
@@ -61,17 +55,13 @@ LED::LED( uint8_t controlPin,
   }
 }
 
-void LED::Initialize()
-{
-  if(!m_initialized)
-  {
-    if(NOT_A_PIN != m_controlPin)
-    {
+void LED::Initialize() {
+  if (!m_initialized) {
+    if (NOT_A_PIN != m_controlPin) {
       pinMode(m_controlPin, OUTPUT);
       digitalWrite(m_controlPin, LOW);
     }
-    if(NOT_A_PIN != m_groundPin)
-    {
+    if (NOT_A_PIN != m_groundPin) {
       pinMode(m_groundPin, OUTPUT);
       digitalWrite(m_groundPin, LOW);
     }
@@ -79,10 +69,8 @@ void LED::Initialize()
   m_initialized = true;
 }
 
-void LED::Update()
-{
-  switch(m_pattern)
-  {
+void LED::Update() {
+  switch (m_pattern) {
     case LEDPattern::BLINK_FAST:
     case LEDPattern::BLINK_SLOW:
       SetFlash();
@@ -98,54 +86,43 @@ void LED::Update()
   }
 }
 
-void LED::On()
-{
+void LED::On() {
   m_illuminate = true;
   Update();
 }
 
-void LED::Off()
-{
+void LED::Off() {
   m_illuminate = false;
   Update();
 }
 
-bool LED::GetIlluminated() const
-{
+bool LED::GetIlluminated() const {
   return m_illuminate;
 }
 
-void LED::SetPattern(LED::LEDPattern newPattern)
-{
+void LED::SetPattern(LED::LEDPattern newPattern) {
   m_pattern = newPattern;
 }
 
-LED::LEDPattern LED::GetPattern() const
-{
+LED::LEDPattern LED::GetPattern() const {
   return m_pattern;
 }
 
-bool LED::IsPWM() const
-{
+bool LED::IsPWM() const {
   return m_pwmEnable;
 }
 
-void LED::SetPulse()
-{
-  if(!m_pwmEnable)
-  {
+void LED::SetPulse() {
+  if (!m_pwmEnable) {
     SetFlash();
     return;
   }
 
-  if(NOT_A_PIN != m_controlPin)
-  {
+  if (NOT_A_PIN != m_controlPin) {
     uint8_t brightness = 0;
-    if(m_illuminate)
-    {
+    if (m_illuminate) {
       uint16_t period = PULSE_SLOW_PERIOD_MS;
-      switch(m_pattern)
-      {
+      switch (m_pattern) {
         case LEDPattern::PULSE_FAST:
           period = PULSE_FAST_PERIOD_MS;
           break;
@@ -159,24 +136,17 @@ void LED::SetPulse()
           return;
           break;
       }
-      brightness = CalcPulseBrightness( PWM_MAX_BRIGHTNESS,
-                                        period,
-                                        true,
-                                        PWM_MIN_BRIGHTNESS );
+      brightness = CalcPulseBrightness(PWM_MAX_BRIGHTNESS, period, true, PWM_MIN_BRIGHTNESS);
     }
     analogWrite(m_controlPin, GAMMA[brightness]);
   }
 }
 
-
-void LED::SetFlash()
-{
-  if(NOT_A_PIN != m_controlPin)
-  {
-    const long now = millis();
+void LED::SetFlash() {
+  if (NOT_A_PIN != m_controlPin) {
+    const int32_t now = millis();
     bool illuminate = m_illuminate;
-    switch(m_pattern)
-    {
+    switch (m_pattern) {
       case LEDPattern::BLINK_FAST:
         illuminate = illuminate && (now / (BLINK_FAST_PERIOD_MS / 2)) % 2;
         break;
@@ -196,40 +166,29 @@ void LED::SetFlash()
   }
 }
 
-void LED::SetSolid()
-{
-  if(NOT_A_PIN != m_controlPin)
-  {
+void LED::SetSolid() {
+  if (NOT_A_PIN != m_controlPin) {
     digitalWrite(m_controlPin, m_illuminate ? HIGH : LOW);
   }
 }
 
-uint8_t LED::CalcPulseBrightness( uint8_t brightness,
-                                  uint16_t period,
-                                  bool smooth,
-                                  uint8_t offBrightness )
-{
+uint8_t LED::CalcPulseBrightness(uint8_t brightness, uint16_t period, bool smooth, uint8_t offBrightness) {
   bool invert = false;
   period /= PERIODDIVISOR;
-  unsigned long condensedNow = millis() / PERIODDIVISOR;
-  if(smooth)
-  {
-    invert = (condensedNow % (2 * period)) !=
-             (condensedNow % period);
+  uint32_t condensedNow = millis() / PERIODDIVISOR;
+  if (smooth) {
+    invert = (condensedNow % (2 * period)) != (condensedNow % period);
   }
   uint16_t curPhase = (condensedNow % period);
-  if(invert)
-  {
+  if (invert) {
     curPhase = period - curPhase;
   }
   uint8_t brightnessRange = brightness - offBrightness;
   uint32_t outBrightness = curPhase * brightnessRange;
   outBrightness /= period;
   outBrightness += offBrightness;
-  if(outBrightness > brightness)
-  {
+  if (outBrightness > brightness) {
     outBrightness = brightness;
   }
   return static_cast<uint8_t>(outBrightness);
 }
-
